Harden ACME bootstrap and renewal

+++ core/base/libs/socrate/socrate/system.py
@@ -80,7 +80,7 @@ def sigterm_handler(_signo, _stack_frame):
     log.critical("Received SIGTERM, terminating.")
     sys.exit(143)
 
-def set_env(required_secrets=[], log_filters=[]):
+def set_env(required_secrets=[], log_filters=[], cleanup_pids=True):
     if log_filters:
         sys.stdout = LogFilter(sys.stdout, log_filters)
         sys.stderr = LogFilter(sys.stderr, log_filters)
@@ -109,7 +109,8 @@ def set_env(required_secrets=[], log_filters=[]):
     for secret in required_secrets:
         os.environ[f'{secret}_KEY'] = hmac.new(bytearray(secret_key, 'utf-8'), bytearray(secret, 'utf-8'), 'sha256').hexdigest()
 
-    os.system(r'find /run -xdev -type f -name \*.pid -print -delete')
+    if cleanup_pids:
+        os.system(r'find /run -xdev -type f -name \*.pid -print -delete')
 
     return {
             key: _coerce_value(os.environ.get(key, value))
